pyproject.toml
[project]
name = "agent-service-toolkit"
version = "0.1.0"
description = "Full toolkit for running an AI agent service built with LangGraph, FastAPI and Streamlit"
readme = "README.md"
authors = [{ name = "Joshua Carroll", email = "carroll.joshk@gmail.com" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Framework :: FastAPI",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

requires-python = ">=3.11"

dependencies = [
    "duckduckgo-search>=7.3.0",
    "fastapi ~=0.115.5",
    "grpcio >=1.68.0",
    "httpx ~=0.27.2",
    "jiter ~=0.8.2",
    "langchain-core ~=0.3.33",
    "langchain-community ~=0.3.16",
    "langchain-openai ~=0.2.9",
    "langchain-anthropic ~= 0.3.0",
    "langchain-google-genai ~=2.0.11",
    "langchain-groq ~=0.2.1",
    "langchain-aws ~=0.2.14",
    "langchain-ollama ~=0.2.3",
    "langgraph ~=0.2.68",
    "langgraph-checkpoint-sqlite ~=2.0.1",
    "langgraph-checkpoint-postgres ~=2.0.13",
    "langsmith ~=0.1.145",
    "numexpr ~=2.10.1",
    "numpy ~=1.26.4; python_version <= '3.12'",
    "numpy ~=2.2.3; python_version >= '3.13'",
    "pandas ~=2.2.3",
    "psycopg[binary,pool] ~=3.2.4",
    "pyarrow >=18.1.0",
    "pydantic ~=2.10.1",
    "pydantic-settings ~=2.6.1",
    "pyowm ~=3.3.0",
    "python-dotenv ~=1.0.1",
    "setuptools ~=75.6.0",
    "streamlit ~=1.40.1",
    "tiktoken >=0.8.0",
    "uvicorn ~=0.32.1",
    "matplotlib>=3.10.1",
    "bs4>=0.0.2",
    "chromadb>=0.6.3",
    "docx2txt>=0.8",
]

[dependency-groups]
dev = [
    "pre-commit",
    "pytest",
    "pytest-cov",
    "pytest-env",
    "pytest-asyncio",
    "ruff",
]

# Group for the minimal dependencies to run just the client and Streamlit app.
# These are also installed in the default dependencies.
# To install run: `uv sync --frozen --only-group client`
client = [
    "httpx~=0.27.2",
    "pydantic ~=2.10.1",
    "python-dotenv ~=1.0.1",
    "streamlit~=1.40.1",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
extend-select = ["I", "U"]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_default_fixture_loop_scope = "function"

[tool.pytest_env]
OPENAI_API_KEY = "sk-fake-openai-key"
